improvements(inter-contexts): Adicionando registro através de outros contextos para manter integridade dos registros

+++ CentralTicket/Contexts/Profile/DTOs/Sale/ReadSaleDTO.cs
@@ -5,7 +5,7 @@ namespace CentralTicket.Contexts.Profile.DTOs.Sale
     public class ReadSaleDTO
     {
         public Guid Id { get; set; }
-        public double TotalValue { get; set; }
+        public decimal TotalValue { get; set; }
         public string Status { get; set; }
         public string PaymentMethod { get; set; }
         public string OrderCode { get; set; }
